SignUp, Login, and JWT Token Generate

+++ src/main/java/id/wiryawan/jwttokenlogin/repository/UserRepository.java
@@ -12,9 +12,9 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsernameOrEmail(String username, String email);
 
-    List<User> findByIdIn(List<Long> userIds);
+    List<User> findByIdIn(List<Integer> userIds);
 
-    Optional<User> findById(Long userId);
+    Optional<User> findById(Integer userId);
 
     Optional<User> findByUsername(String username);
